requirements.txt
numpy==1.23.5
pandas
playwright
pillow
joblib==1.1.0
scikit-learn==1.1.0
imbalanced-learn==0.9.0
streamlit
